feat: 添加模拟用户信息模块

+++ fast-tiny-admin/src/main/java/com/me/fast/tiny/admin/dto/SysArticleSaveRequest.java
@@ -22,7 +22,7 @@ public class SysArticleSaveRequest {
     private Integer categoryId;
 
     @Schema(description = "文章状态")
-    private Byte status;
+    private Integer status;
 
     @Schema(description = "阅读量")
     private Integer viewCount;
